requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
apispec==6.6.0
attrs==23.2.0
bech32==1.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cosmpy==0.9.2
ecdsa==0.18.0
fastapi==0.110.2
frozenlist==1.4.1
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httplib2==0.22.0
idna==3.7
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
msgpack==1.0.8
multidict==6.0.5
packaging==24.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
proxy.py==2.4.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.20.0
pydantic==1.10.15
pydantic_core==2.18.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.2
typing_extensions==4.11.0
uagents==0.11.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.20.0
websockets==10.4
yarl==1.9.4
